backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
snownlp==0.12.3
requests==2.31.0 